fix(nvim): prioritize snippet jump over expand
When jumping through a snippet, if you enter another snippet value, it
expands it instead of doing the jumping. And sometimes you do want that
and other times you don't. So its better to finish the last snippet (by
jumping throught to the end) than starting a new one, so we prioritize
jumping over expanding.