fix(wezterm): hide window decorations unless macos
We like no decorations but on MacOS, no decorations means we cannot
resize the terminal (using Rectangle app or manually). And going
fullscreen mode hides the decorations anyway. All good so far, for macos
or no macos. But going fullscreen hides the system notifications as well
(verified on Linux only, maybe the issue is on MacOS as well).