feat(zsh): enable bundler prefix aliases
Typing `bundle exec rails` and `bundle exec rake` is too much for my
hands. I am a simple man who wants to write `rails` and let shell find
the context of that command.

We had to add `rails` in the `BUNDLED_COMMANDS` array because it is not
there by default unlike `rake` and many other commands.