feat(nvim): add nushell syntax highlighting
Why do we have a build step for installing parser instead of adding "nu"
to ensure_installed in the treesitter config?

First of all, good question. The reason why we have to do it this way is
because the nvim-nu registers the "nu" parser for treesitter. If "nu" is
in ensure_installed table when treesitter is loaded, it will fail
because nvim-nu plugin has not loaded hence not registered the "nu"
parser. One other solution is to add nvim-nu plugin as a dependency of
treesitter plugin but that feels wrong since nvim-nu will be loaded
whenever treesitter is loaded.