fix(tmux): start mapped popups in pwd
The current path is important  for the mapped popups (especially lazygit
and joshuto) hence  we use the path  of the current pane  as we navigate
around the  directories. Before this fix,  it would use the  path of the
session which works for most cases since we don't usually change paths.