fix(nvim): fix piping data to shell commands
When we shifted to nushell, the shell for the neovim also switched. This
is the shell that is used to  run shell commands from the command prompt
(commands that  are prefixed with !).  For some reason, piping  the data
fails when using nushell.

Steps to reproduce:
1. Visually select a line
2. Press colon to enter command mode
3. Enter a shell command (e.g. !echo)

We could have used simple `/bin/sh` here instead of bash but we probably
use some bash magic in our commands without noticing.