feat(nvim): improve git commit message formatting
Using the  par command and  configuring the  format options to  need can
help with better commit messages.

The justify (-j) option  does not work as the text  is typed. There must
be a limitation that prevents this from happening. We can always trigger
manual formatting by visually selecting the paragraph and triggering the
formatting using gq keys.