feat(tmux): remove exit delay on jump windows
`atuin history end` (and `atuin history  start`) is slow. We do not want
to wait for the  time and exit status to be  reported before exiting, it
can be done in the background.

This commit  only moves `atuin history  end` to the background  but same
can be done  for `atuin history start` for improved  startup. We can run
both of these commands at the end.