feat(nvim): add vim-enuch plugin
This plugin has some utility commands that makes life easy. For example,
using the  Rename command, we can  rename the file and  the buffer name.
Without this,  if we  manually update  the filename  (outside of  vim or
using the Oil plugin),  we would still have a buffer  open with old file
name which will raise errors.